+++ x-pack/plugins/ingest_manager/server/services/epm/packages/install.test.ts
@@ -6,32 +6,26 @@
 
 import { ElasticsearchAssetType, Installation, KibanaAssetType } from '../../../types';
 import { SavedObject } from 'src/core/server';
-jest.mock('./install', () => ({
-  ...(jest.requireActual('./install') as {}),
-  bulkInstallPackages: jest.fn(async () => {
-    return [
-      {
-        name: 'blah',
-        assets: [],
-        newVersion: '',
-        oldVersion: '',
-        statusCode: 200,
-      },
-    ];
-  }),
-}));
 
+jest.mock('./install');
+jest.mock('./bulk_install_packages');
 jest.mock('./get', () => ({
   ...(jest.requireActual('./get') as {}),
   getInstallation: jest.fn(async () => {
     return mockInstallation.attributes;
   }),
 }));
-import { getInstallType, ensureInstalledDefaultPackages } from './install';
-
+import { bulkInstallPackages } from './bulk_install_packages';
+const { ensureInstalledDefaultPackages } = jest.requireActual('./install');
 import { savedObjectsClientMock } from 'src/core/server/mocks';
 import { appContextService } from '../../app_context';
 import { createAppContextStartContractMock } from '../../../mocks';
+
+// if we add this assertion, TS will type check the return value
+// and the editor will also know about .mockImplementation, .mock.calls, etc
+const mockedBulkInstallPackages = bulkInstallPackages as jest.MockedFunction<
+  typeof bulkInstallPackages
+>;
 
 const mockInstallation: SavedObject<Installation> = {
   id: 'test-pkg',
@@ -49,22 +43,7 @@ const mockInstallation: SavedObject<Installation> = {
     install_started_at: new Date().toISOString(),
   },
 };
-const mockInstallationUpdateFail: SavedObject<Installation> = {
-  id: 'test-pkg',
-  references: [],
-  type: 'epm-packages',
-  attributes: {
-    id: 'test-pkg',
-    installed_kibana: [{ type: KibanaAssetType.dashboard, id: 'dashboard-1' }],
-    installed_es: [{ type: ElasticsearchAssetType.ingestPipeline, id: 'pipeline' }],
-    es_index_patterns: { pattern: 'pattern-name' },
-    name: 'test packagek',
-    version: '1.0.0',
-    install_status: 'installing',
-    install_version: '1.0.1',
-    install_started_at: new Date().toISOString(),
-  },
-};
+
 describe('install', () => {
   describe('ensureInstalledDefaultPackages', () => {
     beforeEach(async () => {
@@ -74,68 +53,20 @@ describe('install', () => {
       appContextService.stop();
     });
     it('should return an array of Installation objects when successful', async () => {
+      mockedBulkInstallPackages.mockImplementation(async function () {
+        return [
+          {
+            name: 'blah',
+            assets: [],
+            newVersion: '',
+            oldVersion: '',
+            statusCode: 200,
+          },
+        ];
+      });
       const soClient = savedObjectsClientMock.create();
       const resp = await ensureInstalledDefaultPackages(soClient, jest.fn());
       expect(resp).toEqual([mockInstallation.attributes]);
-    });
-  });
-  describe('getInstallType', () => {
-    it('should return correct type when installing and no other version is currently installed', () => {
-      const installTypeInstall = getInstallType({ pkgVersion: '1.0.0', installedPkg: undefined });
-      expect(installTypeInstall).toBe('install');
-
-      // @ts-expect-error can only be 'install' if no installedPkg given
-      expect(installTypeInstall === 'update').toBe(false);
-      // @ts-expect-error can only be 'install' if no installedPkg given
-      expect(installTypeInstall === 'reinstall').toBe(false);
-      // @ts-expect-error can only be 'install' if no installedPkg given
-      expect(installTypeInstall === 'reupdate').toBe(false);
-      // @ts-expect-error can only be 'install' if no installedPkg given
-      expect(installTypeInstall === 'rollback').toBe(false);
-    });
-
-    it('should return correct type when installing the same version', () => {
-      const installTypeReinstall = getInstallType({
-        pkgVersion: '1.0.0',
-        installedPkg: mockInstallation,
-      });
-      expect(installTypeReinstall).toBe('reinstall');
-
-      // @ts-expect-error cannot be 'install' if given installedPkg
-      expect(installTypeReinstall === 'install').toBe(false);
-    });
-
-    it('should return correct type when moving from one version to another', () => {
-      const installTypeUpdate = getInstallType({
-        pkgVersion: '1.0.1',
-        installedPkg: mockInstallation,
-      });
-      expect(installTypeUpdate).toBe('update');
-
-      // @ts-expect-error cannot be 'install' if given installedPkg
-      expect(installTypeUpdate === 'install').toBe(false);
-    });
-
-    it('should return correct type when update fails and trys again', () => {
-      const installTypeReupdate = getInstallType({
-        pkgVersion: '1.0.1',
-        installedPkg: mockInstallationUpdateFail,
-      });
-      expect(installTypeReupdate).toBe('reupdate');
-
-      // @ts-expect-error cannot be 'install' if given installedPkg
-      expect(installTypeReupdate === 'install').toBe(false);
-    });
-
-    it('should return correct type when attempting to rollback from a failed update', () => {
-      const installTypeRollback = getInstallType({
-        pkgVersion: '1.0.0',
-        installedPkg: mockInstallationUpdateFail,
-      });
-      expect(installTypeRollback).toBe('rollback');
-
-      // @ts-expect-error cannot be 'install' if given installedPkg
-      expect(installTypeRollback === 'install').toBe(false);
     });
   });
 });
